knowledge-mcp/pyproject.toml
[project]
name = "knowledge-mcp"
version = "0.1.0"
description = "MCP server for listing and presigning knowledge base documents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.3",
]


